Add game and piece models, w/ serializer & urls
commit is a culmination of yesterday and this morning.
yesterday was not very productive code-wise and was largely spent at the drawing board rethinking the structure of this backend. My initial thinking was in 'thinking in psql' not django, and the translation took some work as these are still new technologies to me.
How to set up the game |--< piece relationship is particularly difficult at this point and is what i am currently trying to implement.

Now fast approaching V1 MVP.
game model is almost finalized, currently working on the nested relationship between the game and it's pieces. added a method to support piece creation in model serializer, see NewModelSerializer

testing with postman up to this point is effective. ( mangos, users, games, but NOT newgame route and serializer )

noticed a bug where the django navigable api doesn't properly format requests with a credentials object. currently unresolved as it is a peripheral.

next steps are to add CRUD routes for pieces individually as a game view,
test them.

extra goodness for testing the built in custom method to 'move' the piece on the patch.

take the info learned from the testing and try to implement in the newgame serializer// request.