After Rebase: fixing the old bugs & gaps
I. AM. A. FOOL. talking about a lesson hard learned. I foolishly removed the mango model and ever place it was referenced without walking back the migrations for the database. The reconciliation was near impossible and the instructor didn't know how to help me. I suggested maybe a git revert... We went along and executed, tried reconciling the merge conflicts, had to restore deleted files and had 'lost' Lots of code in the process.

Eventually figured out I still had the messed up head on main with the code i was missing and manually copied over sections to make this new branch functional.

~ 4 hours later I am committing this branch